requirements.txt
Flask == 3.0.0
PyMySQL == 1.1.0
cryptography == 41.0.7
Flask-SQLAlchemy == 3.1.1
mysqlclient == 2.2.1